requirements.txt
pycryptodome
azure-storage-blob
